Start the test suite for wicked.
There was still an old test for the SQL based driver by Jason Felice
that I included into the new test suite but marked as incomplete. It
has certainly been neglected for a while and doesn't work in its
current form. I lacked the time to rescue it now but maybe somebody
else knows if should be removed because it is just too old or whether
the is a chance to salvage something from it.