Complete basic elements for reStructuredText rendering.
At several points the wiki token parsing and the rst rendering is not
exactly trivial to bridge. So this commit contains a few minor hacks
to allow getting close to what we do with our rst documentation at the
moment.

The rendering engine should be complete enough to allow rendering a
basic README file. This in turn should be sufficient to get
horde-components to fetch the doc files from wiki.horde.org into the
different components.